# perf: Faster singleton SiblingSubgraph construction