--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -81,7 +81,7 @@ repos:
       - id: cargo-clippy
         name: cargo clippy
         description: Run clippy lints with `cargo clippy`.
-        entry: cargo clippy --all-features --workspace -- -D warnings
+        entry: cargo clippy --all-targets --all-features --workspace -- -D warnings
         language: system
         files: \.rs$
         pass_filenames: false