--- a/hugr/benches/bench_main.rs
+++ b/hugr/benches/bench_main.rs
@@ -7,5 +7,6 @@ use criterion::criterion_main;
 
 criterion_main! {
     benchmarks::hugr::benches,
+    benchmarks::subgraph::benches,
     benchmarks::types::benches,
 }