--- a/hugr/benches/benchmarks/mod.rs
+++ b/hugr/benches/benchmarks/mod.rs
@@ -1,2 +1,3 @@
 pub mod hugr;
+pub mod subgraph;
 pub mod types;