@aborgna-q 2024-11-13
Uh, looks like I introduced a breaking change -.-
(`TopoConvexChecker` is no longer `Sync` now that it uses an `OnceCell` internally...)